User land work

+++ Disk/sys/src/gfx/Graphics.cpp
@@ -1,6 +1,5 @@
 #include <sys/lib/gfx/Graphics.h>
-#include <sys/lib/Library.h>
-#include <sys/lib/Common.h>
+#include <sys/lib/api/Library.h>
 
 namespace pmlib
 {
@@ -11,18 +10,18 @@ namespace pmlib
         uint32_t* fb = GetFrameBuffer();
 
         if (data == NULL) { return; }
-        while (x + w > fbw) { w--; }
-        while (y + h > fbh) { h--; }
+        while (x + w > (int)fbw) { w--; }
+        while (y + h > (int)fbh) { h--; }
 
         for (int yy = 0; yy < h; yy++)
         {
             uint8_t* src = (uint8_t*)(data + (yy * w));
             int xx = x;
             if (xx < 0) { xx = 0; }
-            while (xx + w > fbw) { xx--; }
+            while (xx + w > (int)fbw) { xx--; }
             uint32_t real_offset = (xx + ((y + yy) * fbw)) * 4;
             uint8_t* dest = (uint8_t*)((uint32_t)fb + real_offset);
-            if (y + yy >= fbh) { return; }
+            if (y + yy >= (int)fbh) { return; }
             if (y + yy >= 0 && (uint32_t)dest >= (uint32_t)fb && (uint32_t)dest < (uint32_t)((uint32_t)fb + (fbw * fbh * 4)))
             {
                 if (x >= 0) { memcpy(dest, src, w * 4); }
